refactor(engine): consolidate attachment hydration into engine/attachments.ts
delivery.ts, dm.ts, groupDm.ts and message.ts each carried a verbatim copy of
the `AttachmentRow` type and a `fetchAttachmentsBatch` helper that joins
message_attachments -> files and groups by message id. The copies had drifted:
delivery.ts lacked the `position` ordering, and message.ts lacked both the
`files.workspaceId` scope and the ordering.

Extract a single shared helper that always scopes by workspace and orders by
(message_id, position), and point all four modules at it. This removes the
duplication and aligns every read path on the safest, most-correct variant —
delivery payloads and channel/thread message lists now return attachments in a
deterministic position order.

Verified: build, 133 unit tests, lint (0 warnings), 107 live e2e checks, plus a
manual upload->send->list check confirming attachments serialize with correct
fields and position ordering.
